+++ app/team/controller.py
@@ -1,7 +1,7 @@
 from app import db
 from app.auth.controller import get_current_user
 from app.models_db import Team, User
-from app.models_schema import TeamSchema, UserSchema
+from app.models_schema import TeamSchema
 from app.utils import model_list_to_dict_list, user_in_team, get_value_from_payload, user_is_team_owner, check_key
 from flask import Blueprint, jsonify, Response, abort
 
@@ -15,7 +15,8 @@ team_schema_nested = TeamSchema(exclude=("members.teams",))
 @check_key
 @get_current_user()
 def get_teams(current_user):
-    team_objs = db.session.query(Team).filter(Team.members.any(User.id == current_user.id)).all()
+    team_objs = db.session.query(Team).filter(
+        Team.members.any(User.id == current_user.id)).all()
     json_build = model_list_to_dict_list(team_objs, team_schema)
     return jsonify(json_build)
 
@@ -25,6 +26,8 @@ def get_teams(current_user):
 @user_in_team
 def get_team(team_id):
     team = Team.query.get(team_id)
+    if team is None:
+        abort(404, "Team not found")
     json_build = model_list_to_dict_list([team], team_schema_nested)
     return jsonify(json_build[0])
 
@@ -37,7 +40,7 @@ def create_team(current_user):
     description = get_value_from_payload("description")
     sharing = get_value_from_payload("sharing")
     creator = current_user.id
-    team = Team(title, creator, sharing)
+    team = Team(title, description, creator, sharing)
     team.members.append(current_user)
     db.session.add(team)
     db.session.commit()
@@ -49,6 +52,9 @@ def create_team(current_user):
 @user_is_team_owner
 def update_team(team_id):
     team = Team.query.get(team_id)
+
+    if team is None:
+        abort(404, "Team not found")
 
     title = get_value_from_payload("title", optional=True)
     description = get_value_from_payload("description", optional=True)
@@ -76,6 +82,8 @@ def update_team(team_id):
 @user_is_team_owner
 def delete_team(team_id):
     team = Team.query.get(team_id)
+    if team is None:
+        abort(404, "Team not found")
     db.session.delete(team)
     db.session.commit()
     return Response(status=204)
